aams-ai/requirements.txt
flask==3.0.2
flask-cors==4.0.0
mediapipe==0.10.14
opencv-python-headless==4.9.0.80
numpy>=1.24.0
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
scikit-learn==1.4.2